Added the script that runs the experiment
The script shows the UI with fields to enter all the data about the
participant and the task. When the button is clicked the experiment
starts and a graph is shown in a separate window. When a user
clicks on the graph, the timer starts and the ttime elapsed from
the click is shown when the mouse button gets released again.
If the mouse button is down, the coordinates of the mouse pointer
are printed to the console.